+++ backend/app/utils/__init__.py
@@ -1,0 +1,4 @@
+"""Utilities package"""
+from .decorators import admin_required
+
+__all__ = ['admin_required']
