+++ tests/auth/test_user_auth.py
@@ -1,6 +1,7 @@
 """Test user authentication"""
 import pytest
-from app import db, User
+from backend.app import db
+from backend.app.models import User
 
 class TestUserAuth:
     """Test user authentication"""
